+++ test/unit/mysql_rep_failover/main.py
@@ -28,6 +28,94 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        arg_dir_chk
+        arg_require
+        get_val
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.cmdline = None
+        self.args_array = dict()
+        self.opt_req = None
+        self.opt_req2 = True
+        self.dir_perms_chk = None
+        self.dir_perms_chk2 = True
+        self.opt_xor_val = None
+        self.opt_xor_val2 = True
+
+    def arg_dir_chk(self, dir_perms_chk):
+
+        """Method:  arg_dir_chk
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_dir_chk.
+
+        Arguments:
+
+        """
+
+        self.dir_perms_chk = dir_perms_chk
+
+        return self.dir_perms_chk2
+
+    def arg_require(self, opt_req):
+
+        """Method:  arg_require
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_require.
+
+        Arguments:
+
+        """
+
+        self.opt_req = opt_req
+
+        return self.opt_req2
+
+    def arg_xor_dict(self, opt_xor_val):
+
+        """Method:  arg_xor_dict
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_xor_dict.
+
+        Arguments:
+
+        """
+
+        self.opt_xor_val = opt_xor_val
+
+        return self.opt_xor_val2
+
+    def get_val(self, skey, def_val=None):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+
 class ProgramLock(object):
 
     """Class:  ProgramLock
@@ -67,12 +155,12 @@ class UnitTest(unittest.TestCase):
         test_programlock_false
         test_programlock_true
         test_run_program
-        test_arg_dir_chk_crt_false
         test_arg_dir_chk_crt_true
+        test_arg_dir_chk_crt_false
         test_arg_xor_dict_true
         test_arg_xor_dict_false
-        test_arg_require_false
         test_arg_require_true
+        test_arg_require_false
         test_help_false
         test_help_true
 
@@ -88,14 +176,17 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
-        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
+        self.args = ArgParser()
+        self.args2 = ArgParser()
+        self.args.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
+        self.args2.args_array = {
+            "-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("mysql_rep_failover.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_failover.gen_class.ProgramLock")
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_programlock_id(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_programlock_id
@@ -106,12 +197,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array2
+        mock_arg.return_value = self.args2
         mock_help.return_value = False
-        mock_arg.arg_req_xor.return_value = True
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_cond_req.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_rep_failover.main())
@@ -119,7 +206,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_rep_failover.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_failover.gen_class.ProgramLock")
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_programlock_false(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_programlock_false
@@ -130,12 +217,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_req_xor.return_value = True
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_cond_req.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
         mock_lock.side_effect = \
             mysql_rep_failover.gen_class.SingleInstanceException
 
@@ -145,7 +228,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_rep_failover.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_failover.gen_class.ProgramLock")
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_programlock_true(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_programlock_true
@@ -156,12 +239,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_req_xor.return_value = True
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_cond_req.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_rep_failover.main())
@@ -169,7 +248,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_rep_failover.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_failover.gen_class.ProgramLock")
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_run_program(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_run_program
@@ -180,11 +259,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_rep_failover.main())
@@ -192,29 +268,8 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_rep_failover.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_failover.gen_class.ProgramLock")
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
-    def test_arg_dir_chk_crt_false(self, mock_arg, mock_help, mock_lock):
-
-        """Function:  test_arg_dir_chk_crt_false
-
-        Description:  Test arg_dir_chk_crt if returns false.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_lock.return_value = self.proglock
-
-        self.assertFalse(mysql_rep_failover.main())
-
-    @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
-    def test_arg_dir_chk_crt_true(self, mock_arg, mock_help):
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
+    def test_arg_dir_chk_crt_true(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_arg_dir_chk_crt_true
 
@@ -224,16 +279,34 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = True
+        mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_rep_failover.main())
 
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
+    def test_arg_dir_chk_crt_false(self, mock_arg, mock_help):
+
+        """Function:  test_arg_dir_chk_crt_false
+
+        Description:  Test arg_dir_chk_crt if returns false.
+
+        Arguments:
+
+        """
+
+
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
+        mock_help.return_value = False
+
+        self.assertFalse(mysql_rep_failover.main())
+
+    @mock.patch("mysql_rep_failover.gen_libs.help_func")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_arg_xor_dict_true(self, mock_arg, mock_help):
 
         """Function:  test_arg_xor_dict_true
@@ -244,16 +317,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = True
 
         self.assertFalse(mysql_rep_failover.main())
 
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_arg_xor_dict_false(self, mock_arg, mock_help):
 
         """Function:  test_arg_xor_dict_false
@@ -264,34 +336,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        self.args.opt_xor_val2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = False
 
         self.assertFalse(mysql_rep_failover.main())
 
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
-    def test_arg_require_false(self, mock_arg, mock_help):
-
-        """Function:  test_arg_require_false
-
-        Description:  Test arg_require if returns false.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = False
-
-        self.assertFalse(mysql_rep_failover.main())
-
-    @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_arg_require_true(self, mock_arg, mock_help):
 
         """Function:  test_arg_require_true
@@ -302,14 +355,34 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        self.args.opt_xor_val2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
 
         self.assertFalse(mysql_rep_failover.main())
 
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
+    def test_arg_require_false(self, mock_arg, mock_help):
+
+        """Function:  test_arg_require_false
+
+        Description:  Test arg_require if returns false.
+
+        Arguments:
+
+        """
+
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
+        mock_help.return_value = False
+
+        self.assertFalse(mysql_rep_failover.main())
+
+    @mock.patch("mysql_rep_failover.gen_libs.help_func")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_help_false(self, mock_arg, mock_help):
 
         """Function:  test_help_false
@@ -320,14 +393,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
 
         self.assertFalse(mysql_rep_failover.main())
 
     @mock.patch("mysql_rep_failover.gen_libs.help_func")
-    @mock.patch("mysql_rep_failover.arg_parser.arg_parse2")
+    @mock.patch("mysql_rep_failover.gen_class.ArgParser")
     def test_help_true(self, mock_arg, mock_help):
 
         """Function:  test_help_true
@@ -338,7 +412,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.return_value = self.args
         mock_help.return_value = True
 
         self.assertFalse(mysql_rep_failover.main())