+++ test/unit/mysql_rep_failover/promote_designated_slave.py
@@ -28,6 +28,44 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        get_val
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.cmdline = None
+        self.args_array = dict()
+
+    def get_val(self, skey, def_val=None):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+
 class MasterRep(object):
 
     """Class:  MasterRep
@@ -142,6 +180,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.args = ArgParser()
+        self.args2 = ArgParser()
+        self.args.args_array = {"-G": "slave1"}
+        self.args2.args_array = {"-G": "slave0"}
         self.master = MasterRep()
         self.slave1 = SlaveRep("slave1", "20", True)
         self.slave2 = SlaveRep("slave2", "10", True)
@@ -161,8 +203,6 @@ class UnitTest(unittest.TestCase):
         self.slaveorder.append((slv2.exe_gtidset, slv2))
         self.slaveorder.append((slv0.exe_gtidset, slv0))
         self.slaveorder2.append((slv1.exe_gtidset, slv1))
-        self.args_array = {"-G": "slave1"}
-        self.args_array2 = {"-G": "slave0"}
         self.results = \
             "Slaves: ['slave2', 'slave3'] that did not change to new master."
         self.results2 = "Slaves: ['slave2'] that did not change to new master."
@@ -186,8 +226,9 @@ class UnitTest(unittest.TestCase):
 
         mock_master.return_value = self.master
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray, self.args_array), (True, self.results4))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray, self.args), (True, self.results4))
 
     @mock.patch("mysql_rep_failover.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
@@ -206,8 +247,9 @@ class UnitTest(unittest.TestCase):
 
         mock_switch.side_effect = [-1, 0]
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray, self.args_array), (True, self.results2))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray, self.args), (True, self.results2))
 
     @mock.patch("mysql_rep_failover.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
@@ -226,8 +268,9 @@ class UnitTest(unittest.TestCase):
 
         mock_switch.return_value = -1
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray, self.args_array), (True, self.results))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray, self.args), (True, self.results))
 
     @mock.patch("mysql_rep_failover.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
@@ -244,8 +287,9 @@ class UnitTest(unittest.TestCase):
 
         mock_switch.return_value = 0
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray, self.args_array2), (True, self.results3))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray, self.args2), (True, self.results3))
 
     @mock.patch("mysql_rep_failover.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
@@ -264,8 +308,9 @@ class UnitTest(unittest.TestCase):
 
         mock_switch.return_value = 0
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray2, self.args_array), (False, None))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray2, self.args), (False, None))
 
     @mock.patch("mysql_rep_failover.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
@@ -284,8 +329,9 @@ class UnitTest(unittest.TestCase):
 
         mock_switch.return_value = 0
 
-        self.assertEqual(mysql_rep_failover.promote_designated_slave(
-            self.slavearray, self.args_array), (False, None))
+        self.assertEqual(
+            mysql_rep_failover.promote_designated_slave(
+                self.slavearray, self.args), (False, None))
 
 
 if __name__ == "__main__":