+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.4.0] - 2023-08-17
 - Upgraded python-lib to v2.10.1
+- Replace arg_parser.arg_parse2 with gen_class.ArgParser.
 
 ### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
 - main: Removed gen_libs.get_inst call.